pyproject.toml
[project]
name = "aisuite4cn"
version = "1.0.5"
description = "Uniform access layer for LLMs of China"
authors = [{ name = "胡伟", email = "3314672@qq.com" }]
readme = "README.md"
requires-python = ">=3.10, <4"

dependencies = [
    "pydantic (>=2.10.3,<3.0.0)",
    "python-box (>=7.3.2,<8.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (~=0.27.0)",
    "agentscope (>=1.0.0, <2.0.0); extra == 'agentscope'",
    "openai (>=1.82.1); extra == 'openai'",
    "qianfan (>=0.4.12.3); extra == 'qianfan'",
    "llama-index-core (>=0.13.4,<0.14.0); extra == 'llama-index'",
]

[project.optional-dependencies]
moonshot = ["openai"]
ark = ["openai"]
qwen = ["openai"]
qianfan = ["qianfan", "openai"]
hunyuan = ["openai"]
zhipuai = ["openai"]
deepseek = ["openai"]
spark = ["openai"]
ollama = ["openai"]
llama-index = ["llama-index"]
agentscope = ["agentscope"]

all = ["openai", "qianfan", "llama-index", "agentscope"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
llama-index-core = ">=0.13.4,<0.14.0"
openai = "^1.82.1"
agentscope = "^1.0.0"
qianfan = ">=0.4.12.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
llama-index-core = ">=0.13.4,<0.14.0"
openai = "^1.82.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"